feat: 需求开发页面支持编辑预估工期（天）
- 新增 EditableCell 组件支持表格内联编辑
- 预估工期单位从月改为天，支持小数输入（0-365天）
- 后端模型字段从 INTEGER 改为 FLOAT
- 添加数据库迁移脚本
- 新增 18 个单元测试，100% 通过

+++ backend/app/models/requirement.py
@@ -119,8 +119,8 @@ class Requirement(Base, TimestampMixin, TenantMixin):
     target_type: Mapped[str | None] = mapped_column(TargetType)
     target_id: Mapped[int | None] = mapped_column(Integer)
 
-    # Time estimation
-    estimated_duration_months: Mapped[int | None] = mapped_column(Integer)
+    # Time estimation (支持小数天数)
+    estimated_duration_months: Mapped[float | None] = mapped_column(Float)
     complexity_level: Mapped[str | None] = mapped_column(ComplexityLevel)
 
     # Metadata
